requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
docxtpl>=0.16
django-filter>=24.2
drf-spectacular>=0.27
